consolidate logic
