+++ examples/stock_exchange_v1/stock_exchange_v1.py
@@ -1,5 +1,5 @@
 from random import shuffle, random, randint
-from typing import Sequence, Tuple, Dict, Union
+from typing import List, Sequence, Tuple, Dict, Union
 
 import numpy as np
 import pandas as pd
@@ -7,13 +7,10 @@ import pandas as pd
 from agent_exchange.agent import Agent
 from agent_exchange.exchange import Exchange
 
-def dict_repr(self):
-    return self.__dict__.__repr__()
+from stock_exchange_v1_utils import dict_repr, dict_str, round_price
 
-def dict_str(self):
-    return self.__dict__.__str__()
+EXCHANGE_PRECISION = 2 # 2 digits of precision for stock prices
 
-# Make stock exchange
 class StockExchangeV1OrderTypes:
     MARKET="MARKET"
     LIMIT="LIMIT"
@@ -111,15 +108,15 @@ class StockExchangeV1OrderBook:
     def get_bid(self):
         if len(self.bids.index) == 0:
             return self.default_bid
-        return max(self.bids.index)
+        return round_price(max(self.bids.index), EXCHANGE_PRECISION)
 
     def get_ask(self):
         if len(self.asks.index) == 0:
             return self.default_ask
-        return min(self.asks.index)
+        return round_price(min(self.asks.index), EXCHANGE_PRECISION)
 
     def get_spread(self):
-        return self.get_ask() - self.get_bid()
+        return round_price(self.get_ask() - self.get_bid(), EXCHANGE_PRECISION)
 
     def __get_book_from_side(self, book_side: str):
         """Add error checking around accessing the
@@ -357,13 +354,38 @@ class StockExchangeV1(Exchange):
         super().__init__(agents)
         self.order_book: StockExchangeV1OrderBook = StockExchangeV1OrderBook(len(agents), 99.95, 100.05)
         self.order_fills_this_step: Dict = {}
+        
+        # Chronological list of max-bid and min-ask at the end of each step -- can be updated in the `on_step_end` method
+        self.bids = []
+        self.asks = []
+
+    def get_order_book(self) -> StockExchangeV1OrderBook:
+        return self.order_book
+
+    def get_bids(self) -> List[float]:
+        """Return the a list of end-of-step
+        maximum bids, ordered chronologically.
+
+        Returns:
+            List[float]: a list of end-of-step maximum bids, ordered chronologically.
+        """
+        return self.bids
+
+    def get_asks(self) -> List[float]:
+        """Return the a list of end-of-step minimum asks,
+        ordered chronologically.
+
+        Returns:
+            List[float]: a list of end-of-step minimum asks, 
+            ordered chronologically.
+        """
+        return self.asks
 
     def get_exchange_state(self) -> StockExchangeV1OrderBook:
         """We define the state of the exchange
         as the state of the order book.
         """
-        return self.order_book
-
+        return self.get_order_book()
     def get_info(self, agent_index):
         """Relay information about how many shares 
         of the user's previous orders filled.
@@ -372,14 +394,14 @@ class StockExchangeV1(Exchange):
             return self.order_fills_this_step[agent_index]
         else:
             return None
-
+    
     def on_step_end(self):
         """Clean up the state from
         order_fills_this_step
         """
         self.order_fills_this_step = {}
-        # print(self.order_book)
-        # print("======== Ending Step ========\n\n\n\n\n")
+        self.bids.append(self.order_book.get_bid())
+        self.asks.append(self.order_book.get_ask())
         return
 
     def update_exchange_state(self, orders: np.array):
@@ -390,208 +412,11 @@ class StockExchangeV1(Exchange):
         """
 
         agent_and_order = list(zip(range(len(orders)), orders))
-        shuffle(agent_and_order)
+        shuffle(agent_and_order) # process agent actions in random sequence
         for agent_index, order in agent_and_order:
-            # Try to fill the agent's order, and modify
+            # Try to fill the agent's order, and update
             # the order book in the process.
             # print(f"`update_exchange_state`, processing agent {agent_index}")
             self.order_book.update_with_order(agent_index, order, self.order_fills_this_step)
         
         return
-
-
-"""REMOVE
-"""
-class StockAgentV1InternalState:
-    def __init__(self, initial_num_shares: int, initial_capital: float):
-        self.num_shares = [initial_num_shares]
-        self.capital = [initial_capital]
-        
-        # Mappings from price to number of shares
-        self.open_bids = {}
-        self.open_asks = {}
-
-    def on_timestep_passed(self, fill_ticket: Union[type(None), StockExchangeV1FillTicket]):
-        if fill_ticket != None:
-            new_num_shares, new_capital = self.update_with_fill_ticket(fill_ticket)
-        else:
-            new_num_shares, new_capital = self.get_num_shares(), self.get_capital()
-            
-        self.num_shares.append(new_num_shares)
-        self.capital.append(new_capital)
-        
-        
-        
-    def update_with_fill_ticket(self, ticket: StockExchangeV1FillTicket):
-        """Use the fill ticket to update our state variables.
-        """
-        
-        # The state updates for after the update -- this should be modified in this function when capital or num_shares changes
-        new_num_shares = self.get_num_shares()
-        new_capital = self.get_capital()
-
-        # Add new bids
-        for price in ticket.open_bids:
-            StockAgentV1InternalState.increment_or_create(self.open_bids, price, ticket.open_bids[price])
-
-        # Add new asks
-        for price in ticket.open_asks:
-            StockAgentV1InternalState.increment_or_create(self.open_asks, price, ticket.open_asks[price])
-
-        # Remove old bids that were filled in the past time step
-        for price in ticket.closed_bids:
-            shares_bought = ticket.closed_bids[price]
-            new_num_shares += shares_bought
-            new_capital -= price * shares_bought
-            StockAgentV1InternalState.decrement_and_try_delete(self.open_bids, price, shares_bought)
-            
-
-        # Remove old asks that were filled in the past time step
-        for price in ticket.closed_asks:
-            shares_sold = ticket.closed_asks[price]
-            new_num_shares -= shares_sold
-            new_capital += price * shares_sold
-            StockAgentV1InternalState.decrement_and_try_delete(self.open_asks, price, shares_sold)
-            
-        return new_num_shares, new_capital
-
-    def get_num_shares(self):
-        return self.num_shares[-1]
-    
-    def get_capital(self):
-        return self.capital[-1]
-
-    def __repr__(self):
-        return dict_repr(self)
-
-    def __str__(self):
-        return dict_str(self)
-
-    def increment_or_create(D, key, value):
-        """If the key-value pair does not exist yet,
-        then add a new key-value pair with `value`
-        as the value. Otherwise, increment the
-        key's value with `value`.
-        """
-        key = round(key, 2)
-        if key not in D:
-            D[key] = 0
-        D[key] += value
-        if D[key] == 0:
-            del D[key]
-
-    def decrement_and_try_delete(D, key, value):
-        """Decrement a value in a dictionary,
-        and if the new value is 0, then delete
-        the k-v pair from the dictionary.
-        """
-        key = round(key, 2)
-        if key not in D:
-            D[key] = 0
-        D[key] -= value
-
-        if D[key] == 0:
-            del D[key]
-
-class StockAgentV1(Agent):
-    """A base stock trading agent; this agent itself will perform no-ops each iteration.
-    """
-    def __init__(self, initial_num_shares, initial_capital):
-        super().__init__()
-        self.internal_state = StockAgentV1InternalState(initial_num_shares, initial_capital)
-
-    def action_results_update(
-        self, 
-        new_order_book: StockExchangeV1OrderBook, 
-        reward, 
-        done: bool, 
-        fill_ticket: Union[type(None), StockExchangeV1FillTicket]):
-        
-        self.internal_state.on_timestep_passed(fill_ticket)
-        
-    def __repr__(self):
-        return dict_repr(self)
-
-    def __str__(self):
-        return dict_str(self)
-
-    
-class StockAgentV1NaiveMaker(StockAgentV1):
-    """Naive agent that acts as a market maker.
-    If there is a spread, this agent will place
-    an order on the buy side 1/2 of the time, on
-    the sell side 1/2 of the time.
-
-    Note: in V1, there is no way to place multiple
-    orders within the same time step.
-    """
-    def __init__(self, initial_num_shares=1000, initial_capital=100000):
-        super().__init__(initial_num_shares, initial_capital)
-
-    def get_action(self, order_book: StockExchangeV1OrderBook):
-        if random() < .5: # buy just over current bid
-            buy_price = order_book.get_bid() + 0.01 # penny-up on the market-clearing buy price
-            
-            # Randomly buy as little as 0 or as much as we can
-            buy_amount = randint(0, 1000) # limit buy
-
-            order = StockExchangeV1Action(StockExchangeV1OrderTypes.LIMIT, buy_amount, buy_price)
-
-            return order
-        else: # sell just under current ask
-            sell_price = order_book.get_ask() - 0.01
-
-            # Randomly sell as little as 0 and as much as 10
-            sell_amount = randint(0, 1000)
-
-            order = StockExchangeV1Action(StockExchangeV1OrderTypes.LIMIT, -sell_amount, sell_price)
-
-            return order
-
-
-class StockAgentV1NaiveTaker(StockAgentV1):
-    """Naive agent that acts as a liquidity
-    taker. This agent speculates by placing
-    market orders of buy 1/2 of the time and
-    sell 1/2 of the time. Here we ignore
-    constraints on short selling, allowing
-    agents to shorts sell without limit.
-
-    Also, if the taker's order exhausts the
-    order book, then only the portion of their
-    order in the order book gets filled.
-    """
-    def __init__(self, initial_num_shares=1000, initial_capital=100000):
-        super().__init__(initial_num_shares, initial_capital)
-
-    def get_action(self, order_book: StockExchangeV1OrderBook):
-        if random() < .5: # buy
-            expected_buy_price = order_book.get_ask()
-            
-            # Randomly decide on how much to buy
-            max_buy_amount = self.internal_state.get_capital() // expected_buy_price
-            num_shares_to_buy = randint(0, max_buy_amount)
-
-            return StockExchangeV1Action(StockExchangeV1OrderTypes.MARKET, num_shares_to_buy)
-
-        else:
-            expected_sell_price = order_book.get_bid()
-
-            # Randomly decide on how much to sell
-            max_sell_amount = self.internal_state.get_num_shares()
-            num_shares_to_sell = randint(0, max_sell_amount)
-
-            return StockExchangeV1Action(StockExchangeV1OrderTypes.MARKET, -num_shares_to_sell)
-if __name__ == "__main__":
-    # Run a simple 
-    NMAKER, NTAKER, NSTEPS = 10, 1, 10
-    agents = [StockAgentV1NaiveMaker() for _ in range(NMAKER)]
-    agents += [StockAgentV1NaiveTaker() for _ in range(NTAKER)]
-    exchange = StockExchangeV1(agents)
-
-    exchange.simulate_steps(NSTEPS)
-    # print(exchange.order_book)
-
-
-
-
